requirements.txt
django
psycopg2-binary
django-tables2
python-dotenv
djangorestframework